% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coast_dist.R
\name{coast_dist}
\alias{coast_dist}
\title{Attach distance to coast to an ocean cube}
\usage{
coast_dist(x, coast)
}
\arguments{
\item{x}{An \verb{<ocean_cube>} object.}

\item{coast}{Either an \code{sf}/\code{sfc} object or a path readable by \code{sf::st_read()}.}
}
\value{
An \verb{<ocean_cube>} with \code{x$dc} as a \verb{[lon, lat]} matrix in nautical miles.
}
\description{
Attach distance to coast to an ocean cube
}
