% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-inspect.R
\name{print.ocean_cube_inspection}
\alias{print.ocean_cube_inspection}
\title{Print an ocean cube inspection}
\usage{
\method{print}{ocean_cube_inspection}(x, ...)
}
\arguments{
\item{x}{An \code{ocean_cube_inspection} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an ocean cube inspection
}
