% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-validate.R
\name{print.oceancube_validation}
\alias{print.oceancube_validation}
\title{Print an ocean cube validation report}
\usage{
\method{print}{oceancube_validation}(x, ...)
}
\arguments{
\item{x}{An \code{oceancube_validation} report.}

\item{...}{Additional arguments, currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an ocean cube validation report
}
