% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_index.R
\name{annual_index}
\alias{annual_index}
\title{Compute annual indicators from an ocean cube}
\usage{
annual_index(x, threshold_pos = NULL, threshold_neg = NULL)
}
\arguments{
\item{x}{An \verb{<ocean_cube>} object.}

\item{threshold_pos}{Optional positive threshold. If \code{NULL}, fraction above zero is returned.}

\item{threshold_neg}{Optional negative threshold. If \code{NULL}, fraction below zero is returned.}
}
\value{
A data frame of annual indicators by variable and depth layer.
}
\description{
Compute annual indicators from an ocean cube
}
