% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anom_diff.R
\name{anom_diff}
\alias{anom_diff}
\title{Compute absolute anomalies from monthly climatology}
\usage{
anom_diff(x, clim)
}
\arguments{
\item{x}{An \verb{<ocean_cube>} object.}

\item{clim}{An \verb{<ocean_clim>} object created by \code{\link[=clim_month]{clim_month()}}.}
}
\value{
An \verb{<ocean_cube>} with anomaly values \code{x - climatology}.
}
\description{
Compute absolute anomalies from monthly climatology
}
