% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anom_diff.R
\name{anom_z}
\alias{anom_z}
\title{Compute standardized monthly anomalies}
\usage{
anom_z(x, clim)
}
\arguments{
\item{x}{An \verb{<ocean_cube>} object.}

\item{clim}{An \verb{<ocean_clim>} object created by \code{\link[=clim_month]{clim_month()}}.}
}
\value{
An \verb{<ocean_cube>} with standardized anomaly values \code{(x - climatology) / sd_clim}.
}
\description{
Compute standardized monthly anomalies
}
