% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_month.R
\name{clim_month}
\alias{clim_month}
\title{Compute monthly climatology from an ocean cube}
\usage{
clim_month(x, period = NULL)
}
\arguments{
\item{x}{An \verb{<ocean_cube>} object.}

\item{period}{Optional base period \code{c(start, end)}.}
}
\value{
An object of class \verb{<ocean_clim>}.
}
\description{
Compute monthly climatology from an ocean cube
}
