% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_setup.R
\name{cm_connect}
\alias{cm_connect}
\title{Connect to the Copernicus Marine Python module}
\usage{
cm_connect(
  env = "oceancube-copernicus",
  required = TRUE,
  module = "copernicusmarine",
  verbose = TRUE
)
}
\arguments{
\item{env}{Virtual environment name.}

\item{required}{Logical. Should activation fail if the environment is missing?}

\item{module}{Python module name.}

\item{verbose}{Logical. Print progress messages?}
}
\value{
Imported Python module.
}
\description{
Connect to the Copernicus Marine Python module
}
