% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_mask.R
\name{crop_stock}
\alias{crop_stock}
\title{Apply a stock mask to an ocean cube}
\usage{
crop_stock(x, mask)
}
\arguments{
\item{x}{An \verb{<ocean_cube>} object.}

\item{mask}{An \verb{<ocean_mask>} object.}
}
\value{
A masked \verb{<ocean_cube>} object.
}
\description{
Apply a stock mask to an ocean cube
}
