% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_mean.R
\name{layer_mean}
\alias{layer_mean}
\title{Average an ocean cube over vertical layers}
\usage{
layer_mean(x, depth)
}
\arguments{
\item{x}{An \verb{<ocean_cube>} object with at least two depth levels.}

\item{depth}{Numeric vector. Either \code{c(zmin, zmax)} for one layer or a sequence
of breaks, e.g. \code{c(0, 10, 25, 50, 100)}.}
}
\value{
An \verb{<ocean_cube>} with the depth dimension replaced by vertical layers.
}
\description{
Average an ocean cube over vertical layers
}
