% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anom_diff.R
\name{signal_noise}
\alias{signal_noise}
\title{Compute signal-to-noise anomaly magnitude}
\usage{
signal_noise(x, clim, signed = FALSE)
}
\arguments{
\item{x}{An ocean_cube object.}

\item{clim}{An ocean_clim object created by \code{\link[=clim_month]{clim_month()}} or \code{\link[=clim_day]{clim_day()}}.}

\item{signed}{Logical. If FALSE, returns absolute signal-to-noise magnitude.}
}
\value{
An ocean_anom object with signal-to-noise values.
}
\description{
Compute signal-to-noise anomaly magnitude
}
