% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anom_diff.R
\name{signal_noise}
\alias{signal_noise}
\title{Compute signal-to-noise anomaly magnitude}
\usage{
signal_noise(x, clim, signed = FALSE)
}
\arguments{
\item{x}{An \verb{<ocean_cube>} object.}

\item{clim}{An \verb{<ocean_clim>} object created by \code{clim_month()} or \code{clim_day()}.}

\item{signed}{Logical. If \code{FALSE}, returns absolute signal-to-noise magnitude.}
}
\value{
An \verb{<ocean_cube>} with signal-to-noise values.
}
\description{
Compute signal-to-noise anomaly magnitude
}
