% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_month.R
\name{to_month}
\alias{to_month}
\title{Aggregate an ocean cube to monthly resolution}
\usage{
to_month(x, fun = mean)
}
\arguments{
\item{x}{An \verb{<ocean_cube>} object.}

\item{fun}{Aggregation function. Defaults to mean.}
}
\value{
An \verb{<ocean_cube>} aggregated to one value per month.
}
\description{
Aggregate an ocean cube to monthly resolution
}
